refactor(proxy): Move binary code into `api` crate
To benefit from all linting we move all code from `api/src/main.rs` to
`api/src/lib.rs`.

Fixes #1052
