+++ proxy/api/src/config.rs
@@ -1,6 +1,6 @@
 //! Configuration vital to the setup and alteration of the application.
 
-use std::{env, ffi, io, path};
+use std::{env, io, path};
 
 use directories::ProjectDirs;
 
@@ -50,23 +50,4 @@ pub fn proxy_path() -> Result<path::PathBuf, Error> {
         .parent()
         .ok_or_else(|| Error::MissingExePath(exe_path.clone()))?
         .to_owned())
-}
-
-/// Include the path to the current running binary in the global system PATH.
-///
-/// # Errors
-///
-///   * Could not determine path to the proxy binary.
-///   * Could not join paths.
-pub fn default_bin_path() -> Result<ffi::OsString, Error> {
-    let proxy_path = proxy_path()?;
-
-    let paths = std::env::var_os("PATH").map_or(vec![proxy_path.to_owned()], |path| {
-        let mut paths = std::env::split_paths(&path).collect::<Vec<_>>();
-        paths.push(proxy_path.to_owned());
-        paths.reverse();
-        paths
-    });
-
-    Ok(std::env::join_paths(paths)?)
 }
